test: move AddQueryParam routing IT from test-react-router to test-default
First step of migrating the React routing suite into test-default (the
default React home). AddQueryParamView + IT move verbatim (IT rewritten to
JUnit 6 + @TestFor) and are deleted from test-react-router in the same
commit, so the test runs in exactly one place. test-vaadin-router keeps
its own vaadin-router version (parallel engine suites, no dedup).
